short fixes and refacto of divMedias functions
