+++ svelte/components/Nft/NftGrid.svelte
@@ -2,8 +2,7 @@
   import type { NftType } from "lib/ktypes";
 
   import { nftName, nftOpenSeaUrl, addressSame, getNetwork } from "lib/kconfig";
-
-  import { nftGetImageLink } from "lib/knft-get-metadata";
+  import { divMedia } from "helpers/mediasDisplay";
 
   /////////////////////////////////////////////////
   //  <NftGrid {chainId} {tokenID} {account} {nft} {platform}? />
@@ -17,6 +16,7 @@
   export let index: number;
 
   export let platform: string = "dapp";
+
   ///////////////////////////////////////////////////////////////////////////////
   const displayNftSolo = (evt: Event, clickedNftTokenID: string) => {
     if (!(evt.target as HTMLInputElement).classList.contains("btn")) {
@@ -24,49 +24,9 @@
       tokenID = clickedNftTokenID;
     }
   };
-  ///////////////////////////////////////////////////////////////////////////////
-  const divMediaImage = (src: string, height?: number) => {
-    const heightString = height ? `height="${height}"` : "";
-    return `<img alt="link" src=${src} ${heightString}/>`;
-  };
 
-  const divMediaVideo = (src: string, small = true) => {
-    let video: string;
-    if (small) {
-      video = '<video preload="metadata" style="border-radius: initial;">';
-    } else {
-      video =
-        '<video autoplay="true"  controls="" controlslist="nodownload" loop="" playsinline="" preload="metadata" style="border-radius: initial;">';
-    }
-    video += `<source src="${src}" type="video/mp4"></video>`;
-    return video;
-  };
-
-  const divMedia = (nft: NftType, index: number, small = false) => {
-    const mediaContentType = nft.contentType?.split("/");
-    const mediaType = mediaContentType[0] || "image";
-
-    const mediaSrc = nftGetImageLink(nft);
-    let div: string = "";
-    if (small) {
-      div += `<div id="media-small-${index}" class="media media-grid media-${mediaType}">`;
-    } else {
-      div += `<div id="media-full-${index}" class="media media-grid media-${mediaType}">`;
-    }
-    div += `<a href=".${/*normalizedSoloNftUrl(chainId, nft)*/ ""}">`;
-    if (mediaType == "video") {
-      div += divMediaVideo(mediaSrc, small);
-    } else if (mediaType == "image") {
-      div += divMediaImage(mediaSrc);
-    } else {
-      div += '<div class="media-text"></div>';
-    }
-    div += "</a></div>";
-
-    // console.log("divMedia div", div);
-    return div;
-  };
   console.log("nft", nft);
+  ///////////////////////////////////////////////////////////////////////////////
 </script>
 
 <div class="col col-xs-12 col-sm-6 col-md-4 col-lg-3">
