+++ svelte/helpers/mediasDisplay.ts
@@ -1,0 +1,78 @@
+import type { NftType } from "lib/ktypes";
+import { nftGetImageLink } from "lib/knft-get-metadata";
+
+const divMediaImage = (src: string, height?: number) => {
+  const heightString = height ? `height="${height}"` : "";
+  return `<img alt="link" src=${src} ${heightString}/>`;
+};
+
+const divMediaVideo = (src: string, small = true) => {
+  let video: string;
+  if (small) {
+    video = '<video preload="metadata" style="border-radius: initial;">';
+  } else {
+    video =
+      '<video autoplay="true"  controls="" controlslist="nodownload" loop="" playsinline="" preload="metadata" style="border-radius: initial;">';
+  }
+  video += `<source src="${src}" type="video/mp4"></video>`;
+  return video;
+};
+
+const divMedia = (nft: NftType, index: number, small = false, display: string = "grid") => {
+  const mediaContentType = nft.contentType?.split("/");
+  const mediaType = mediaContentType[0] || "image";
+
+  const mediaSrc = nftGetImageLink(nft);
+  let div: string = "";
+  if (small) {
+    if ("list" === display) {
+      div = `<div id="media-small-${index}" class="media media-small media-${mediaType}">`;
+    } else {
+      div += `<div id="media-small-${index}" class="media media-grid media-${mediaType}">`;
+    }
+  } else {
+    if ("list" === display) {
+      div = `<div id="media-full-${index}" class="media media-${mediaType}">`;
+    } else {
+      div += `<div id="media-full-${index}" class="media media-grid media-${mediaType}">`;
+    }
+  }
+  div += `<a href=".${/*normalizedSoloNftUrl(chainId, nft)*/ ""}">`;
+  if (mediaType == "video") {
+    div += divMediaVideo(mediaSrc, small);
+  } else if (mediaType == "image") {
+    div += divMediaImage(mediaSrc);
+  } else {
+    div += '<div class="media-text"></div>';
+  }
+  div += "</a></div>";
+
+  // console.log("divMedia div", div);
+  return div;
+};
+
+export { divMedia };
+
+// const divMedia = (_nft: NftType, index: number, small = false) => {
+//   const mediaContentType = _nft.contentType?.split("/");
+//   const mediaType = mediaContentType?.[0] || "image";
+
+//   const mediaSrc = nftGetImageLink(_nft);
+//   let div: string;
+//   if (small) {
+//     div = `<div id="media-small-${index}" class="media media-small media-${mediaType}">`;
+//   } else {
+//     div = `<div id="media-full-${index}" class="media media-${mediaType}">`;
+//   }
+//   if (mediaType == "video") {
+//     div += divMediaVideo(mediaSrc, small);
+//   } else if (mediaType == "image") {
+//     div += divMediaImage(mediaSrc);
+//   } else {
+//     div += '<div class="media-text"></div>';
+//   }
+//   div += "</div>";
+
+//   // console.log("divMedia div", div);
+//   return div;
+// };
